--- src/VirtoCommerce.ElasticAppSearch.Core/Models/Api/EngineType.cs
@@ -1,7 +1,0 @@
-namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api;
-
-public enum EngineType
-{
-    Default,
-    Meta
-}
