+++ src/VirtoCommerce.ElasticAppSearch.Data/Models/EngineType.cs
@@ -1,0 +1,7 @@
+namespace VirtoCommerce.ElasticAppSearch.Data.Models;
+
+public enum EngineType
+{
+    Default,
+    Meta
+}
