+++ src/VirtoCommerce.ElasticAppSearch.Data/Models/FieldType.cs
@@ -1,4 +1,4 @@
-namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api;
+namespace VirtoCommerce.ElasticAppSearch.Data.Models;
 
 public enum FieldType
 {
