+++ src/VirtoCommerce.ElasticAppSearch.Data/Models/Result.cs
@@ -1,4 +1,4 @@
-namespace VirtoCommerce.ElasticAppSearch.Core.Models.Api;
+namespace VirtoCommerce.ElasticAppSearch.Data.Models;
 
 public record Result
 {
