+++ src/VirtoCommerce.ElasticAppSearch.Data/VirtoCommerce.ElasticAppSearch.Data.csproj
@@ -17,7 +17,6 @@
     <Folder Include="Caching\" />
     <Folder Include="ExportImport\" />
     <Folder Include="Handlers\" />
-    <Folder Include="Models\" />
     <Folder Include="Extensions\" />
   </ItemGroup>
